Xlsxir v1.4.1 Xlsxir.ConvertTime

Converts a time formatted as a decimal number that represents the fraction of the day in char_list form, to a tiem formatted in Erlang :erlang.time() format (i.e. {hour, minute, second}).

Functions

from_charlist(charlist)

Given a charlist in the form of a fraction of the day, return the time formatted as :erlang.time().

Note that the float representation of the input must be a number between 0 (inclusive) and 1 (exclusive).

Parameters

charlist - Character list in the form of a fractional number (i.e. 0.25)

Example

iex> Xlsxir.ConvertTime.from_charlist('0.25')
{6, 0, 0}