Xlsxir v0.0.1 Xlsxir.Format

Receives parsed excel worksheet data, formats the cell values and returns the data in either a list or a map depending on the option chosen.

Functions

cell_map(sheet, strings)

Formats parsed excel worksheet into a map of cell/value pairs.

Parameters

sheet - map of xml worksheet data from Excel file that was parsed via Xlsxir.Parse.worksheet/2

strings - list of sharedStrings.xml data from Excel file that was parsed via Xlsxir.Parse.shared_strings/1

Example

%{ A1: value_of_cell, B1: value_of_cell, ...}

convert_char_number(number)

format_cell_value(list, strings)

Uses attributes from xml to format cell value.

Parameters

list - list containing attribute and value of column from xml file

strings - list of strings from sharedStrings.xml file

Example

iex> Xlsxir.Format.format_cell_value([nil, '1'], ["A", "B"])
1
iex> Xlsxir.Format.format_cell_value(['s', '1'], ["A", "B"])
"B"

prepare_output(worksheet, shared_strings, option)

Main function of Format module. Receives the parsed excel worksheet data, the parsed excel string data and the chosen format via option which defaults to :rows.

row_list(sheet, strings)

Formats parsed excel worksheet into a list of lists containing cell values by row.

Parameters

sheet - map of xml worksheet data from Excel file that was parsed via Xlsxir.Parse.worksheet/2

strings - list of sharedStrings.xml from Excel file that was parsed via Xlsxir.Parse.shared_strings/1

Example

[[row_1_values], [row_2_values], ...]