Xlsxir v1.1.0 Xlsxir.Index

An Agent process named Index which holds state of an index. Provides functions to create the process, increment the index by 1, retrieve the current index and ultimately kill the process.

Functions

delete()

Deletes Index Agent process

get()

Returns current value of Index Agent process

increment_1()

Increments active Index Agent process by 1.

new()

Initiates a new Index Agent process with a value of 0.