Xlsxir v0.0.1 Xlsxir.Parse

Receives Excel xml data via the extract_xml function of the Unzip module and parses it.

Functions

shared_strings(arg)

Receives Excel string data in xml format, parses it and returns the strings in the form of a list.

Parameters

path - file path of a .xlsx file type in string format

Example

An example file named test.xlsx located in ./test/test_data containing the following:

cell ‘A1’ -> “string one”

cell ‘B1’ -> “string two”

cell ‘C1’ -> integer of 10

cell ‘D1’ -> formula of =4*5

cell ‘E1’ -> date of 1/1/2016 or Excel date serial of 42370

iex> Xlsxir.Parse.shared_strings({:ok, "./test/test_data/test.xlsx"})
["string one", "string two"]

worksheet(arg, index)

Receives the xlsx worksheet at position index in xml format, parses the data and returns required elements in the form of a keyword list:

[[row_1_cell_1: ['attribute', 'value'], ...], [row_2_cell_1: ['attribute', 'value'], ...], ...]

Parameters

path - file path of a .xlsx file type in string format

index - index of worksheet from within the Excel workbook to be parsed, starting with 0

Example

An example file named test.xlsx located in ./test/test_data containing the following in worksheet at index 0:

cell ‘A1’ -> “string one”

cell ‘B1’ -> “string two”

cell ‘C1’ -> integer of 10

cell ‘D1’ -> formula of =4*5

cell ‘E1’ -> date of 1/1/2016 or Excel date serial of 42370

iex> Xlsxir.Parse.worksheet({:ok, "./test/test_data/test.xlsx"}, 0)
[[A1: ['s', '0'], B1: ['s', '1'], C1: [nil, '10'], D1: [nil, '20'], E1: ['1', '42370']]]