Xlsxir v1.3.3 Xlsxir.ParseWorksheet

Holds the SAX event instructions for parsing worksheet data via Xlsxir.SaxParser.parse/2

Functions

sax_event_handler(arg1, state)

Sax event utilized by Xlsxir.SaxParser.parse/2. Takes a pattern and the current state of a struct and recursivly parses the worksheet XML file, ultimately sending a keyword list of cell references and their assocated values to the Xlsxir.Worksheet module which contains an ETS process that was started by Xlsxir.SaxParser.parse/2.

Parameters

arg1 - the XML pattern of the event to match upon

state - the state of the %Xlsxir.ParseWorksheet{} struct which temporarily holds applicable data of the current row being parsed

Example

Each entry in the list created consists of a list containing a cell reference string and the associated value (i.e. [["A1", "string one"], ...]).