Xlsxir v1.0.0 Xlsxir.SharedString

An Erlang Term Storage (ETS) process named :sharedstrings which holds state for data parsed from sharedStrings.xml. Provides functions to create the process, add and retreive data, and ultimately kill the process.

Functions

add_shared_string(shared_string, index)

alive?()

delete()

get_at(index)

new()