Xlsxir v1.1.0 Xlsxir.Style

An Erlang Term Storage (ETS) process named :styles which holds state for data parsed from styles.xml. Provides functions to create the process, add and retreive data, and ultimately kill the process. Also includes a temporary Agent process named NumFmtIds which is utilized during the parsing of the styles.xml file to temporarily hold state of each NumFmtId contained within the file.

Functions

add_id(num_fmt_id)

Adds a numFmtId to the NumFmtIds Agent process.

add_style(style, index)

Stores a style type at a given index in the ETS process.

alive?()

Validates whether the ETS process is active, returning true or false.

delete()

Deletes the ETS process from memory.

delete_id()

Deletes NumFmtIds Agent process.

get_at(index)

Returns a style type at a given index of the ETS process.

get_id()

Returns a list of numFmtIds stored in the NumFmtIds Agent process.

new()

Initializes new ETS process with [:set, :protected, :named_table] options. Additionally, initiates an Agent process to temporarily hold numFmtIds for Xlsxir.ParseStyle.