Xlsxir v1.3.0 Xlsxir.TableId

An Agent process named TableId which temporarily holds the table id of an ETS process. Provides functions to create the process, assign a table id, retrieve the current table id and ultimately kill the process.

Functions

alive?()

assign_id(id)

Assigns an ETS table id to the TableId Agent process.

delete()

Deletes TableId Agent process

get()

Returns current value of TableId Agent process

new()

Initiates a new TableId Agent process with a value of 0.