Xlsxir v1.3.0 Xlsxir.Timer

An Agent process named Time which holds state for time elapsed since execution. Provides functions to start and stop the process, with the stop function returning the time elapsed as a list (i.e. [hour, minute, second, microsecond]).

Functions

start()

Initiates a new Time Agent process. Records current time via :erlang.timestamp and saves it to the Agent process.

stop()

Records current time via :erlang.timestamp and compares it to the timestamp held by the Time Agent process to determin the amount of time elapsed. Returns the time elapsed in the format of [hour, minute, second, microsecond].