Xlsxir v1.5.1 Xlsxir.Worksheet

An Erlang Term Storage (ETS) process named :worksheet which holds state for data parsed from sheet#{n}.xml at index n. Provides functions to create the process, add and retreive data, and ultimately kill the process.

Functions

add_row(row, index, id \\ :worksheet)

Stores a row at a given index in the ETS process.

alive?(id \\ :worksheet)

Validates whether the ETS process is active, returning true or false.

delete(id \\ :worksheet)

Deletes the ETS process from memory.

get_at(row_num, id \\ :worksheet)

Returns a row at a given index of the ETS process.

new()

Initializes new ETS process with [:set, :protected, :named_table] options.

new_multi()

Initializes new ETS process with [:set, :protected] options and assigns the associated table id to Xlsxir.TableId.