Xlsxir v1.3.5 Xlsxir.ConvertDate

Converts an ISO 8601 date format serial number, in char_list format, to a date formatted in Erlang :calendar.date() type format (i.e. {year, month, day}).

Functions

convert_char_number(number)

Converts extracted number in char_list format to either integer or float.

from_serial(serial)

Receives an ISO 8601 date format serial number and returns a date formatted in Erlang :calendar.date() type format.

Parameters

serial - ISO 8601 date format serial in char_list format (i.e. 4/30/75 as ‘27514’)

Example

iex> Xlsxir.ConvertDate.from_serial('27514')
{1975, 4, 30}