Xlsxir v1.6.3 Xlsxir.ConvertDateTime

Converts a datetime formatted as a decimal number that represents the fraction of the day in char_list form, to an elixir naive datetime

Functions

from_charlist(charlist)

Given a charlist in the form of a serial date float representing fraction of the day return a naive datetime

Parameters

charlist - Character list in the form of a date serial with a fractional number (i.e. 41261.6013888889)

Example

iex> Xlsxir.ConvertDateTime.from_charlist('41261.6013888889')
~N[2012-12-18 14:26:00]

from_float(n)